LPS-71451 Move to the proper bundles and packages
