+++ modules/apps/web-experience/asset/asset-api/bnd.bnd
@@ -2,6 +2,7 @@ Bundle-Name: Liferay Asset API
 Bundle-SymbolicName: com.liferay.asset.api
 Bundle-Version: 1.0.0
 Export-Package:\
+	com.liferay.asset.constants,\
 	com.liferay.asset.util,\
 	com.liferay.asset.util.comparator
 Liferay-Releng-Module-Group-Description:
