+++ modules/apps/web-experience/asset/asset-tags-admin-web/build.gradle
@@ -10,8 +10,7 @@ dependencies {
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "javax.servlet.jsp", name: "javax.servlet.jsp-api", version: "2.3.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:web-experience:asset:asset-api")
 	provided project(":apps:web-experience:asset:asset-service")
 	provided project(":apps:web-experience:asset:asset-taglib")
-	provided project(":apps:web-experience:export-import:export-import-api")
-	provided project(":apps:web-experience:export-import:export-import-service")
 }
