+++ modules/apps/web-experience/asset/asset-tags-admin-web/src/main/java/com/liferay/asset/tags/admin/web/internal/portlet/route/AssetTagsAdminFriendlyURLMapper.java
@@ -14,7 +14,7 @@
 
 package com.liferay.asset.tags.admin.web.internal.portlet.route;
 
-import com.liferay.asset.tags.admin.web.internal.constants.AssetTagsAdminPortletKeys;
+import com.liferay.asset.constants.AssetTagsAdminPortletKeys;
 import com.liferay.portal.kernel.portlet.DefaultFriendlyURLMapper;
 import com.liferay.portal.kernel.portlet.FriendlyURLMapper;
